Added Gerber Files for UTM 1.0
Added the Gerber Files, Schematics and BOM for the UTM 1.0 PCB.

This Design is no longer recommended and is not advised for commercial use. UTM 2.0 will replace this in the near Future. Please don't mess with this if you don't know much about Electronics. The used LED Driver on this PCB needs some adjustments that could if not done properly increase the Eyestrain and in worst case add permanent Eye damage to you.
All you do with this is up to your own responsibility. I will NOT be made accountable for damage that results of misuse of those Resources.